refactor(github): extract remote repository identity cache to fix max-lines
